[Platinum V] Title: 히스토그램에서 가장 큰 직사각형, Time: 32 ms, Memory: 2940 KB -BaekjoonHub

+++ C++17/백준/Platinum/6549. 히스토그램에서 가장 큰 직사각형/히스토그램에서 가장 큰 직사각형.cc
@@ -5,20 +5,20 @@
 
 using namespace std;
 
-ll histogram[100000];
-ll histogramSize = 0;
+int histogram[100000];
+int histogramSize = 0;
 
 //인자 : 시작idx, 끝idx, 너비유지 최대높이
 //반환값 : 다음 계산 여부
 ll largestRectangle() {
-	stack<ll> s;
+	stack<int> s;
 	ll maxArea = 0;
 
-	for (ll i = 0; i <= histogramSize; i++) {
-		ll curHeight = (i == histogramSize) ? 0 : histogram[i];
+	for (int i = 0; i <= histogramSize; i++) {
+		int curHeight = (i == histogramSize) ? 0 : histogram[i];
 
 		while (!s.empty() && histogram[s.top()] > curHeight) {
-			ll h = histogram[s.top()];
+			int h = histogram[s.top()];
 			s.pop();
 			ll width = s.empty() ? i : (i - s.top() - 1);
 			maxArea = max(maxArea, h * width);
@@ -37,7 +37,7 @@ int main() {
 		if (histogramSize == 0) break;
 
 		//히스토그램 입력
-		for (ll i = 0; i < histogramSize; i++) {
+		for (int i = 0; i < histogramSize; i++) {
 			cin >> histogram[i];
 		}
 		cout << largestRectangle() << '\n';
